Refactor SDL2 DLL paths in config-example.cmake for consistency
Enhance base3D example with current working directory debug output

Update pointerLock example to use NoxEngine includes

Add README.md for build scripts detailing cleanup operations

Implement remove_cpp_from_headers.cmake script for header organization

Remove unnecessary output in Console.cpp message function

Correct include paths in IniSet and Actor files to use string.hpp

Refactor BufferGeometry and Shader files to utilize string.hpp

Change Camera class member types from double to float for consistency

Refactor PointerLockControls to use deltaTime consistently

Update Renderer class to improve type consistency and cleanup

Add engine.export.hpp for DLL export/import handling

Include necessary headers in engine.hpp and engine.typedef.hpp

Enhance Color.hpp with glm vector includes for color manipulation

Introduce string.cpp and string.hpp for string utility functions

+++ src/core/Camera/Camera.hpp
@@ -15,7 +15,7 @@
 #include "core/engine.typedef.hpp"
 #include "core/MatricesMVP/Matrices.hpp"
 #include "core/Actor/Movable.hpp"
-#include "utils/utils.hpp"
+#include "utils/string.hpp"
 
 namespace NoxEngine {
 
@@ -78,7 +78,7 @@ namespace NoxEngine {
             }
 
             bool m_needsUpdate;
-            double m_near, m_far;
+            float m_near, m_far;
             V3D m_target, m_verticalAxis;
             Matrices m_matrix;
     };
